+++ docs/make.jl
@@ -19,8 +19,8 @@ gallery, postprocess_cb, gallery_assets = makedemos("gallery")
 
 makedocs(;
     modules=[AlgebraOfGraphics],
-    authors="Pietro Vertechi <pietro.vertechi@veos.digital>",
-    repo="https://github.com/JuliaPlots/AlgebraOfGraphics.jl/blob/{commit}{path}#{line}",
+    authors="Pietro Vertechi",
+    repo="https://github.com/MakieOrg/AlgebraOfGraphics.jl/blob/{commit}{path}#{line}",
     sitename="Algebra of Graphics",
     format=Documenter.HTML(;
         prettyurls=get(ENV, "CI", "false") == "true",
@@ -57,6 +57,6 @@ makedocs(;
 postprocess_cb() # redirect url for DemoCards generated files
 
 deploydocs(;
-    repo="github.com/JuliaPlots/AlgebraOfGraphics.jl",
+    repo="github.com/MakieOrg/AlgebraOfGraphics.jl",
     push_preview=true,
 )