+++ src/transformations/linear.jl
@@ -24,8 +24,7 @@ function (l::LinearAnalysis)(input::ProcessedLayer)
         x̂ = range(extrema(x)..., length=l.npoints)
         pred = GLM.predict(lin_model, add_intercept_column(x̂); interval, l.level)
         return if !isnothing(interval)
-            # TODO: fix this in GLM https://github.com/JuliaStats/GLM.jl/pull/467
-            ŷ, lower, upper = map(vec, pred) # GLM prediction returns matrices
+            ŷ, lower, upper = pred
             (x̂, ŷ), (; lower, upper)
         else
             ŷ = pred
@@ -49,5 +48,7 @@ Use `interval = nothing` to only compute the line fit, without any uncertainty e
 By default, this analysis errors on singular (collinear) data. To avoid that,
 it is possible to set `dropcollinear=true`.
 `npoints` is the number of points used by Makie to draw the shaded band.
+
+Weighted data is supported via the keyword `weights` (passed to `mapping`).
 """
 linear(; options...) = transformation(LinearAnalysis(; options...))